LPS-56029 Move AutoLogin related classes to portal-service/com.liferay.portal.kernel.security.auto.login
